+++ packages/backend/src/services/SemanticLayerService/SemanticLayerService.ts
@@ -4,25 +4,31 @@ import {
     MissingConfigError,
     SemanticLayerField,
     SemanticLayerQuery,
-    SemanticLayerResultRow,
+    SemanticLayerQueryPayload,
     SemanticLayerView,
     SessionUser,
 } from '@lightdash/common';
 import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
+import { S3Client } from '../../clients/Aws/s3';
 import CubeClient from '../../clients/cube/CubeClient';
 import DbtCloudGraphqlClient from '../../clients/dbtCloud/DbtCloudGraphqlClient';
 import { LightdashConfig } from '../../config/parseConfig';
+import Logger from '../../logging/logger';
+import { DownloadFileModel } from '../../models/DownloadFileModel';
 import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
+import { SchedulerClient } from '../../scheduler/SchedulerClient';
 import { BaseService } from '../BaseService';
 
 type SearchServiceArguments = {
     lightdashConfig: LightdashConfig;
     analytics: LightdashAnalytics;
     projectModel: ProjectModel;
-
+    downloadFileModel: DownloadFileModel;
     // Clients
+    schedulerClient: SchedulerClient;
     cubeClient: CubeClient;
     dbtCloudClient: DbtCloudGraphqlClient;
+    s3Client: S3Client;
 };
 
 export class SemanticLayerService extends BaseService {
@@ -32,19 +38,28 @@ export class SemanticLayerService extends BaseService {
 
     private readonly projectModel: ProjectModel;
 
+    private readonly downloadFileModel: DownloadFileModel;
+
+    private readonly schedulerClient: SchedulerClient;
+
     // Clients
     private readonly cubeClient: CubeClient;
 
     private readonly dbtCloudClient: DbtCloudGraphqlClient;
+
+    private readonly s3Client: S3Client;
 
     constructor(args: SearchServiceArguments) {
         super();
         this.analytics = args.analytics;
         this.lightdashConfig = args.lightdashConfig;
         this.projectModel = args.projectModel;
+        this.downloadFileModel = args.downloadFileModel;
+        this.schedulerClient = args.schedulerClient;
         // Clients
         this.cubeClient = args.cubeClient;
         this.dbtCloudClient = args.dbtCloudClient;
+        this.s3Client = args.s3Client;
     }
 
     private async checkCanViewProject(user: SessionUser, projectUuid: string) {
@@ -117,22 +132,60 @@ export class SemanticLayerService extends BaseService {
     async getFields(
         user: SessionUser,
         projectUuid: string,
-        table: string,
+        view: string,
+        selectedFields: Pick<
+            SemanticLayerQuery,
+            'dimensions' | 'timeDimensions' | 'metrics'
+        >,
     ): Promise<SemanticLayerField[]> {
         await this.checkCanViewProject(user, projectUuid);
         const client = await this.getSemanticLayerClient(projectUuid);
-        return client.getFields(table);
+        return client.getFields(view, selectedFields);
     }
 
-    async getResults(
+    async getStreamingResults(
         user: SessionUser,
         projectUuid: string,
         query: SemanticLayerQuery,
-    ): Promise<SemanticLayerResultRow[]> {
+    ) {
         await this.checkCanViewProject(user, projectUuid);
+        await this.getSemanticLayerClient(projectUuid); // Check if client is available
+
+        const jobId = await this.schedulerClient.semanticLayerStreamingResults({
+            projectUuid,
+            userUuid: user.userUuid,
+            query,
+            context: 'semanticViewer',
+        });
+
+        return { jobId };
+    }
+
+    async streamQueryIntoFile({
+        userUuid,
+        projectUuid,
+        query,
+        context,
+    }: SemanticLayerQueryPayload): Promise<{
+        fileUrl: string;
+    }> {
+        // TODO add analytics
+        Logger.debug(`Streaming query into file for project ${projectUuid}`);
         const client = await this.getSemanticLayerClient(projectUuid);
 
-        return client.getResults(query);
+        const fileUrl = await this.downloadFileModel.streamFunction(
+            this.s3Client,
+        )(
+            `${this.lightdashConfig.siteUrl}/api/v2/projects/${projectUuid}/semantic-layer/results`,
+            async (writer) => {
+                await client.streamResults(projectUuid, query, async (rows) => {
+                    rows.forEach(writer);
+                });
+            },
+            this.s3Client,
+        );
+
+        return { fileUrl };
     }
 
     async getSql(