+++ packages/backend/src/services/ServiceRepository.ts
@@ -672,8 +672,12 @@ export class ServiceRepository
                     lightdashConfig: this.context.lightdashConfig,
                     analytics: this.context.lightdashAnalytics,
                     projectModel: this.models.getProjectModel(),
+                    downloadFileModel: this.models.getDownloadFileModel(),
+
+                    schedulerClient: this.clients.getSchedulerClient(),
                     cubeClient: this.clients.getCubeClient(),
                     dbtCloudClient: this.clients.getDbtCloudGraphqlClient(),
+                    s3Client: this.clients.getS3Client(),
                 }),
         );
     }